Add temporary mapping of old ticker names in stockrow.py
This is definitely not ideal, but stockrow has not updated FB -> META ticker symbols yet. For now, I'm just adding a temporary mapping as who knows how long that may be.

This isn't a scalable approach and will fail if/when they finally rename it (at which point this can just be removed).

However this was the best quick-and-dirty solution I could think of.