First step at reviving the repro off MSNMoney.
Morningstar, StockRow, and YahooFinance are all largely dead for
scraping now. However MSNMoney still has a lot of the data we need
available.

I moved all the old code to src/Obsolete for now (for reference or
incase we can find a way to revive it, such as extracting an API key
from a request).

We now use MSNMoney for pretty much everything. There's one Yahoo
Finance API still referenced for the current stock price.

Any data I still need to get is denoted with -999999.